Unify markdown formatting in message views
